+++ src/components/map/popups/MapPopupPilot.vue
@@ -413,12 +413,17 @@ function handlePointerDrag() {
 
 onMounted(() => {
     watch(() => pilot.value.airport, async icao => {
-        if (airportInfo.value?.icao === icao) return;
+        try {
+            if (airportInfo.value?.icao === icao) return;
 
-        if (icao) {
-            props.overlay.data.airport = await $fetch<VatsimAirportInfo>(`/api/data/vatsim/airport/${ icao }/info`);
+            if (icao) {
+                props.overlay.data.airport = await $fetch<VatsimAirportInfo>(`/api/data/vatsim/airport/${ icao }/info`);
+            }
+            else props.overlay.data.airport = undefined;
         }
-        else props.overlay.data.airport = undefined;
+        catch {
+
+        }
     }, {
         immediate: true,
     });